replace dynamic plugin import in provider/auth with static import, fix cycle checker
