+++ packages/opencode/src/server/server.ts
@@ -13,7 +13,7 @@ import { TuiRoutes } from "./routes/tui"
 import { Instance } from "../project/instance"
 import { runPromise as runPromiseVcs, Vcs } from "../project/vcs"
 import { Agent } from "../agent/agent"
-import { Skill } from "../skill/skill"
+import { Skill } from "../skill"
 import { Auth } from "../auth"
 import { Flag } from "../flag/flag"
 import { Command } from "../command"
